refactor(LottoRepository): LottoWrapper의 의미가 불분명, LottoRepository로 네이밍 변경
